Move di to separate folder
