+++ lib/backend/fs_interface/data/helper/fsops.dart
@@ -2,19 +2,6 @@ import 'dart:io';
 
 import 'path_op_string.dart';
 
-/// Returns a String path list under the given [path].
-List<String> getUnder<T extends FileSystemEntity>(
-  final String path,
-) {
-  final dir = Directory(path);
-  if (!dir.existsSync()) {
-    return [];
-  }
-  final res =
-      dir.listSync().whereType<T>().map((final event) => event.path).toList();
-  return res;
-}
-
 /// Returns a String path list under the given [path], synchronously.
 List<String> getUnderSync<T extends FileSystemEntity>(
   final String path,
@@ -47,25 +34,6 @@ String? findPreviewFileInString(
       continue;
     }
     final ext = element.pExtension;
-    for (final previewExt in _previewExtensions) {
-      if (ext.pEquals(previewExt)) {
-        return element;
-      }
-    }
-  }
-  return null;
-}
-
-(String, int)? findPreviewFileInStringTuple(
-  final List<(String, int)> paths, {
-  final String name = 'preview',
-}) {
-  for (final element in paths) {
-    final filename = element.$1.pBNameWoExt;
-    if (!filename.pEquals(name)) {
-      continue;
-    }
-    final ext = element.$1.pExtension;
     for (final previewExt in _previewExtensions) {
       if (ext.pEquals(previewExt)) {
         return element;
