+++ lib/backend/mod_writer/data/mod_writer.dart
@@ -44,7 +44,7 @@ Future<String> _getNonCollidingName(
   final ModCategory category,
   final String destDirName,
 ) async {
-  final enabledFormDirNames = getUnder<Directory>(category.path)
+  final enabledFormDirNames = getUnderSync<Directory>(category.path)
       .map((final e) => e.pDisabledForm.pBasename)
       .toSet();
   var counter = 0;
