+++ lib/ui/di/fs_interface.dart
@@ -1,9 +1,9 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../storage/di/storage.dart';
-import '../../storage/domain/usecase/ini_editor_arg.dart';
-import '../data/fs_interface.dart';
-import '../domain/fs_interface.dart';
+import '../../backend/fs_interface/data/repo/fs_interface.dart';
+import '../../backend/fs_interface/domain/repo/fs_interface.dart';
+import 'storage.dart';
+import '../../backend/storage/domain/usecase/ini_editor_arg.dart';
 
 part 'fs_interface.g.dart';
