+++ lib/ui/route/category/category.dart
@@ -7,8 +7,8 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../backend/fs_interface/di/fs_interface.dart';
-import '../../../backend/fs_interface/di/fs_watcher.dart';
+import '../../di/fs_interface.dart';
+import '../../di/fs_watcher.dart';
 import '../../../backend/fs_interface/domain/usecase/file_system.dart';
 import '../../../backend/structure/entity/mod.dart';
 import '../../../backend/structure/entity/mod_category.dart';
