+++ lib/ui/route/category/mod_card.dart
@@ -11,10 +11,10 @@ import 'package:pasteboard/pasteboard.dart';
 import '../../../backend/akasha/data/akasha.dart';
 import '../../../backend/fs_interface/data/helper/mod_switcher.dart';
 import '../../../backend/fs_interface/data/helper/path_op_string.dart';
-import '../../../backend/fs_interface/di/fs_interface.dart';
-import '../../../backend/fs_interface/di/mod_card.dart';
+import '../../di/fs_interface.dart';
+import '../../di/mod_card.dart';
 import '../../../backend/mod_writer/data/mod_writer.dart';
-import '../../../backend/storage/di/app_state.dart';
+import '../../di/app_state.dart';
 import '../../../backend/structure/entity/ini.dart';
 import '../../../backend/structure/entity/mod.dart';
 import '../../util/display_infobar.dart';
