+++ lib/ui/route/home_shell.dart
@@ -12,15 +12,15 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:http/http.dart' as http;
 import 'package:window_manager/window_manager.dart';
 
-import '../../backend/app_arg/di/exe_arg.dart';
-import '../../backend/app_version/di/app_version.dart';
+import '../di/exe_arg.dart';
+import '../di/app_version.dart';
 import '../../backend/app_version/domain/github.dart';
 import '../../backend/fs_interface/data/helper/path_op_string.dart';
-import '../../backend/fs_interface/di/fs_interface.dart';
-import '../../backend/fs_interface/di/fs_watcher.dart';
-import '../../backend/storage/di/app_state.dart';
-import '../../backend/storage/di/storage.dart';
+import '../di/app_state.dart';
+import '../di/storage.dart';
 import '../../backend/structure/entity/mod_category.dart';
+import '../di/fs_interface.dart';
+import '../di/fs_watcher.dart';
 import '../route_names.dart';
 import '../util/display_infobar.dart';
 import '../util/open_url.dart';
