+++ lib/ui/route/setting.dart
@@ -9,10 +9,10 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../backend/app_version/di/app_version.dart';
-import '../../backend/fs_interface/di/fs_interface.dart';
+import '../di/app_version.dart';
+import '../di/fs_interface.dart';
 import '../../backend/fs_interface/domain/entity/setting_data.dart';
-import '../../backend/storage/di/app_state.dart';
+import '../di/app_state.dart';
 import '../../backend/storage/domain/entity/game_config.dart';
 import '../../backend/storage/domain/usecase/card_color.dart';
 import '../route_names.dart';
