+++ lib/ui/widget/game_selector.dart
@@ -4,7 +4,7 @@ import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../backend/storage/di/app_state.dart';
+import '../di/app_state.dart';
 import '../util/display_infobar.dart';
 
 /// Game selector widget.
