+++ lib/ui/widget/preset_control.dart
@@ -5,7 +5,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../backend/storage/di/preset.dart';
+import '../di/preset.dart';
 import '../../backend/structure/entity/mod_category.dart';
 
 /// A widget that provides a control for presets.
